requirements.txt
pip~=24.0
asgiref==3.8.1
dj-database-url==2.2.0
Django==5.0.4
django-cors-headers==4.3.1
django-filter==24.2
django-heroku==0.3.1
django-js-asset==2.2.0
django-mptt==0.16.0
djangorestframework==3.15.1
drf-yasg==1.21.7
gunicorn==22.0.0
inflection==0.5.1
packaging==24.0
pillow==10.3.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pytz==2024.1
PyYAML==6.0.1
setuptools==70.0.0
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
whitenoise==6.7.0
djangorestframework-simplejwt==5.3.1
